requirements.txt
apache-airflow[postgres,celery,redis]==2.8.4
pandas==2.1.4
pyarrow==14.0.2
boto3==1.33.13
pymongo==4.6.2
iam-rolesanywhere-session==2.1.0
